+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Fix indexing: `Config.from_yaml_str(...)[key]` is now a Config object if it's a mapping
+
 ## v0.10.3 (2025-01-19)
 
 - Don't fail but warn, when we suspect that un unquoted string might actually be an object in config files.